Move all routing into its own folder
