+++ services/backend/src/controllers/courses-controller.js
@@ -2,8 +2,8 @@ const _ = require('lodash');
 const { dynamoUtils } = require('@logan/aws');
 const { v4: uuid } = require('uuid');
 const Promise = require('bluebird');
-const requestValidator = require('../utils/request-validator');
-const { NotFoundError, ValidationError } = require('../utils/errors');
+const requestValidator = require('../../utils/request-validator');
+const { NotFoundError, ValidationError } = require('../../utils/errors');
 const assignmentsController = require('./assignments-controller');
 const tasksController = require('./tasks-controller');
