+++ services/backend/src/controllers/ping.js
