+++ services/backend/src/controllers/sections-controller.js
@@ -4,8 +4,8 @@ const {
     dateUtils: { dayjs, constants: dateConstants },
 } = require('@logan/core');
 const { v4: uuid } = require('uuid');
-const requestValidator = require('../utils/request-validator');
-const { NotFoundError } = require('../utils/errors');
+const requestValidator = require('../../utils/request-validator');
+const { NotFoundError } = require('../../utils/errors');
 
 function fromDbFormat(db) {
     return {
