+++ services/backend/src/controllers/tasks-controller.js
@@ -1,8 +1,8 @@
 const _ = require('lodash');
 const { dynamoUtils } = require('@logan/aws');
 const { v4: uuid } = require('uuid');
-const requestValidator = require('../utils/request-validator');
-const { NotFoundError } = require('../utils/errors');
+const requestValidator = require('../../utils/request-validator');
+const { NotFoundError } = require('../../utils/errors');
 const { remindersForEntity } = require('./reminders-controller');
 
 function fromDbFormat(db) {
