+++ services/backend/src/controllers/terms-controller.js
@@ -8,8 +8,8 @@ const {
 } = require('@logan/core');
 const { v4: uuid } = require('uuid');
 const Promise = require('bluebird');
-const requestValidator = require('../utils/request-validator');
-const { NotFoundError } = require('../utils/errors');
+const requestValidator = require('../../utils/request-validator');
+const { NotFoundError } = require('../../utils/errors');
 const coursesController = require('./courses-controller');
 
 function fromDbFormat(db) {
